pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dinnote"
version = "0.1.0"
description = "Audio denoising, VAD, speaker diarization, and transcription pipeline using Demucs, Silero VAD, pyannote, and Whisper"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "demucs>=4.0.0",
    "torchcodec",
    "torch",
    "torchaudio",
    "pydub",
    "packaging",
    "pyannote.audio>=3.1.0",
    "openai-whisper",
    "pyyaml",
]

[project.scripts]
dinnote = "dinnote.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
dinnote = ["config.yaml", "vocab.txt"]
